+++ apps/storybook/.storybook/manager.js
@@ -1,0 +1,6 @@
+import {addons} from '@storybook/manager-api'
+import theme from '../src/theme'
+
+addons.setConfig({
+  theme,
+})