+++ packages/design-tokens/scripts/build-tokens.js
@@ -204,6 +204,8 @@ const colorModeFormat = require('../src/formats/color-mode-attributes')
     `tokens/functional/components/footer/colors.json`,
     `tokens/functional/components/action-menu/colors.js`,
     `tokens/functional/components/grid/colors.json`,
+    `tokens/functional/components/logosuite/colors.json`,
+    `tokens/functional/components/timeline/colors.json`,
   ]
 
   for (const path of filesForColorModes) {