+++ packages/react/src/Button/Button.module.css.d.ts
@@ -3,7 +3,6 @@ declare const styles: {
   readonly "Button__text": string;
   readonly "Button__icon-visual": string;
   readonly "Button__trailing-visual": string;
-  readonly "Button__trailing-visual--arrow": string;
   readonly "Button__leading-visual": string;
   readonly "Button--label": string;
   readonly "Button-label--disabled": string;
@@ -17,6 +16,7 @@ declare const styles: {
   readonly "Button--label-secondary": string;
   readonly "Button--size-medium": string;
   readonly "Button--size-large": string;
+  readonly "Button--block": string;
 };
 export = styles;