+++ packages/react/src/Heading/Heading.visual.spec.ts
@@ -14,8 +14,22 @@ test.describe('Visual Comparison: Heading', () => {
     expect(await page.screenshot()).toMatchSnapshot()
   })
 
-  test('Heading / Scale', async ({page}) => {
+  test('Heading / Playground', async ({page}) => {
+    await page.goto('http://localhost:6006/iframe.html?args=&id=components-heading--playground&viewMode=story')
+
+    await page.waitForTimeout(500)
+    expect(await page.screenshot()).toMatchSnapshot()
+  })
+
+  test('Heading / Scale (sizes)', async ({page}) => {
     await page.goto('http://localhost:6006/iframe.html?args=&id=components-heading--scale&viewMode=story')
+
+    await page.waitForTimeout(500)
+    expect(await page.screenshot()).toMatchSnapshot()
+  })
+
+  test('Heading / Levels', async ({page}) => {
+    await page.goto('http://localhost:6006/iframe.html?args=&id=components-heading--levels&viewMode=story')
 
     await page.waitForTimeout(500)
     expect(await page.screenshot()).toMatchSnapshot()