+++ packages/react/src/Pillar/Pillar.tsx
@@ -72,7 +72,7 @@ function PillarIcon({icon: Icon, className, color = defaultPillarIconColor, ...r
   return (
     <span className={clsx(styles.Pillar__icon, styles[`Pillar__icon--color-${color}`], className)} {...rest}>
       {typeof Icon === 'function' ? (
-        <Icon size={24} />
+        <Icon size="medium" />
       ) : (
         React.isValidElement(Icon) &&
         React.cloneElement(Icon as React.ReactElement, {
@@ -90,7 +90,7 @@ type PillarHeadingProps = BaseProps<HTMLHeadingElement> & {
 } & HeadingProps
 
 const PillarHeading = forwardRef<HTMLHeadingElement, PillarHeadingProps>(
-  ({children, as = 'h3', size = '6', className, ...rest}, ref) => {
+  ({children, as = 'h3', size = 'subhead-large', className, ...rest}, ref) => {
     return (
       <Heading
         size={size}
@@ -113,7 +113,7 @@ const PillarDescription = forwardRef<HTMLParagraphElement, PillarDescriptionProp
       <Text
         variant="muted"
         ref={ref}
-        size="300"
+        size="200"
         as="p"
         className={clsx(styles.Pillar__description, className)}
         {...rest}