+++ packages/react/src/VideoPlayer/VideoPlayer.features.stories.tsx
@@ -1,0 +1,24 @@
+import React from 'react'
+import {Meta} from '@storybook/react'
+import posterImage from '../fixtures/images/example-poster.png'
+
+import {VideoPlayer} from '.'
+
+export default {
+  title: 'Components/VideoPlayer/Features',
+  component: VideoPlayer,
+} as Meta<typeof VideoPlayer>
+
+export const WithPoster = () => (
+  <VideoPlayer poster={posterImage} title="GitHub media player">
+    <VideoPlayer.Source src="/example.mp4" type="video/mp4" />
+    <VideoPlayer.Track src="/example.vtt" />
+  </VideoPlayer>
+)
+
+export const WithoutBranding = () => (
+  <VideoPlayer branding={false} title="GitHub media player">
+    <VideoPlayer.Source src="/example.mp4" type="video/mp4" />
+    <VideoPlayer.Track src="/example.vtt" />
+  </VideoPlayer>
+)